+++ sys/kern/src/arch/arm_m.rs
@@ -76,6 +76,7 @@ use core::sync::atomic::{AtomicBool, AtomicPtr, AtomicU32, Ordering};
 use zerocopy::FromBytes;
 
 use crate::atomic::AtomicExt;
+use crate::descs::RegionAttributes;
 use crate::startup::with_task_table;
 use crate::task;
 use crate::time::Timestamp;
@@ -344,12 +345,12 @@ pub fn apply_memory_protection(task: &task::Task) {
             | (1 << 4)  // honor the region number
             | region.base;
         let ratts = region.attributes;
-        let xn = !ratts.contains(abi::RegionAttributes::EXECUTE);
+        let xn = !ratts.contains(RegionAttributes::EXECUTE);
         // These AP encodings are chosen such that we never deny *privileged*
         // code (i.e. us) access to the memory.
-        let ap = if ratts.contains(abi::RegionAttributes::WRITE) {
+        let ap = if ratts.contains(RegionAttributes::WRITE) {
             0b011
-        } else if ratts.contains(abi::RegionAttributes::READ) {
+        } else if ratts.contains(RegionAttributes::READ) {
             0b010
         } else {
             0b001
@@ -358,10 +359,10 @@ pub fn apply_memory_protection(task: &task::Task) {
         // shareability (with other cores or masters). See table B3-13 in the
         // ARMv7-M ARM. (Settings are identical on v6-M but the sharability and
         // TEX bits tend to be ignored.)
-        let (tex, scb) = if ratts.contains(abi::RegionAttributes::DEVICE) {
+        let (tex, scb) = if ratts.contains(RegionAttributes::DEVICE) {
             // Device memory.
             (0b000, 0b001)
-        } else if ratts.contains(abi::RegionAttributes::DMA) {
+        } else if ratts.contains(RegionAttributes::DMA) {
             // Conservative settings for normal memory assuming that DMA might
             // be a problem:
             // - Outer and inner non-cacheable.
@@ -447,29 +448,28 @@ pub fn apply_memory_protection(task: &task::Task) {
         let rnr = i as u32;
 
         let ratts = region.attributes;
-        let xn = !ratts.contains(abi::RegionAttributes::EXECUTE);
+        let xn = !ratts.contains(RegionAttributes::EXECUTE);
         // ARMv8m has less granularity than ARMv7m for privilege
         // vs non-privilege so there's no way to say that privilege
         // can be read write but non-privilge can only be read only
         // This _should_ be okay?
-        let ap = if ratts.contains(abi::RegionAttributes::WRITE) {
+        let ap = if ratts.contains(RegionAttributes::WRITE) {
             0b01 // RW by any privilege level
-        } else if ratts.contains(abi::RegionAttributes::READ) {
+        } else if ratts.contains(RegionAttributes::READ) {
             0b11 // Read only by any privilege level
         } else {
             0b00 // RW by privilege code only
         };
 
-        let (mair, sh) = if ratts.contains(abi::RegionAttributes::DEVICE) {
+        let (mair, sh) = if ratts.contains(RegionAttributes::DEVICE) {
             // Most restrictive: device memory, outer shared.
             (0b00000000, 0b10)
-        } else if ratts.contains(abi::RegionAttributes::DMA) {
+        } else if ratts.contains(RegionAttributes::DMA) {
             // Outer/inner non-cacheable, outer shared.
             (0b01000100, 0b10)
         } else {
-            let rw = u32::from(ratts.contains(abi::RegionAttributes::READ))
-                << 1
-                | u32::from(ratts.contains(abi::RegionAttributes::WRITE));
+            let rw = u32::from(ratts.contains(RegionAttributes::READ)) << 1
+                | u32::from(ratts.contains(RegionAttributes::WRITE));
             // write-back transient, not shared
             (0b0100_0100 | rw | rw << 4, 0b00)
         };
