+++ sys/kern/src/lib.rs
@@ -38,6 +38,7 @@
 pub mod arch;
 
 pub mod atomic;
+mod descs;
 pub mod err;
 pub mod header;
 pub mod kipc;
